requirements.txt
tensorflow>=2.0.0
pillow>=8.0.0
numpy>=1.19.0
scipy>=1.5.0
matplotlib>=3.3.0
